+++ libs/eslint-plugin-internal/src/index.mjs
@@ -3,6 +3,7 @@ import exampleScreenDefaultRule from './example-screen-default/index.mjs';
 import figmaConnectImportsPackageMatchRule from './figma-connect-imports-package-match/index.mjs';
 import figmaConnectImportsRequiredRule from './figma-connect-imports-required/index.mjs';
 import noDeprecatedJsdocRule from './no-deprecated-jsdoc/index.mjs';
+import { processor as noTypescriptInJsxCodeblockProcessor } from './no-typescript-in-jsx-codeblock/index.mjs';
 import safelySpreadPropsRule from './safely-spread-props/index.mjs';
 
 const plugin = {
@@ -14,6 +15,9 @@ const plugin = {
     'no-deprecated-jsdoc': noDeprecatedJsdocRule,
     'figma-connect-imports-required': figmaConnectImportsRequiredRule,
     'figma-connect-imports-package-match': figmaConnectImportsPackageMatchRule,
+  },
+  processors: {
+    mdx: noTypescriptInJsxCodeblockProcessor,
   },
   configs: {},
 };